rename TestResultsQueue
